% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_checkers.R
\name{validate_file}
\alias{validate_file}
\title{Validate and prepare financial file}
\usage{
validate_file(file)
}
\arguments{
\item{file}{A filepath to the csv to be processed}
}
\value{
A validated and processed csv file.
}
\description{
Validate and prepare financial file
}
