% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_file_validity.R
\name{validate_file}
\alias{validate_file}
\title{Validate and prepare aux-bourses-citoyens financial file}
\usage{
validate_file(filePath)
}
\arguments{
\item{file}{A filepath to the csv to be processed}
}
\value{
A validated and augmented dataframe
}
\description{
Validate and prepare aux-bourses-citoyens financial file
}
